refactor(core): normalize newer component file names
