+++ packages/core/src/core/index.ts
@@ -19,8 +19,8 @@ export * from './ui/error-dialog/data';
 export * from './ui/error-dialog/error-dialog-i18n';
 export * from './ui/fullscreen-button/core';
 export * from './ui/fullscreen-button/data';
-export * from './ui/gesture/gesture-core';
-export * from './ui/hotkey/hotkey-core';
+export * from './ui/gesture/core';
+export * from './ui/hotkey/core';
 export * from './ui/indicator/indicator-labels';
 export * from './ui/indicator/indicator-lifecycle';
 export * from './ui/input-action/input-action';
@@ -76,8 +76,8 @@ export * from './ui/time-slider/time-slider-chapters/data';
 export * from './ui/time-slider/time-slider-chapters/types';
 export * from './ui/time-slider/time-slider-chapters/vars';
 export * from './ui/time-slider/vars';
-export * from './ui/title/title-core';
-export * from './ui/title/title-data-attrs';
+export * from './ui/title/core';
+export * from './ui/title/data';
 export * from './ui/tooltip/core';
 export * from './ui/tooltip/data';
 export * from './ui/tooltip/group-core';
