+++ packages/core/src/core/ui/gesture/gesture-component.ts
@@ -1,5 +1,5 @@
 import { defineComponent } from 'vjsc/components';
 
-import type { GestureProps } from './gesture-core';
+import type { GestureProps } from './core';
 
 export default defineComponent<GestureProps>({ name: 'Gesture' });
