+++ packages/core/src/core/ui/hotkey/hotkey-component.ts
@@ -1,5 +1,5 @@
 import { defineComponent } from 'vjsc/components';
 
-import type { HotkeyProps } from './hotkey-core';
+import type { HotkeyProps } from './core';
 
 export default defineComponent<HotkeyProps>({ name: 'Hotkey' });
