+++ packages/core/src/core/ui/title/data.ts
@@ -1,5 +1,5 @@
 import type { StateAttrMap } from '../types';
-import type { TitleState } from './title-core';
+import type { TitleState } from './core';
 
 export const TitleDataAttrs = {
   /** Present when the element is hidden because no title is available. */
