+++ packages/core/src/dom/hotkey/hotkey.ts
@@ -1,6 +1,6 @@
 import { isMacOS } from '@videojs/utils/dom';
 
-import type { HotkeyProps } from '../../core/ui/hotkey/hotkey-core';
+import type { HotkeyProps } from '../../core/ui/hotkey/core';
 import { HotkeyCoordinator } from './coordinator';
 
 export type HotkeyModifierKey = 'shift' | 'ctrl' | 'alt' | 'meta';
